fix: load sprite on player ready if configured at player level
Regression introduced in 4c7a616e0f75b359129f835967780047c41e0007:

When the player is installed on a video tag with sources the ready
callback which initializes the plugin is fired after loadstart.
Therefore the image fails to load when it is entirely configured in the
plugin object as a player option.