fix: do not try to show thumbnails before video metadata is loaded
Video duration is required to determine which thumbnail is shown.
Therefore `loadedmetadata` is the correct player event to determine
whether the plugin is ready show thumbnails.

As the options, except for `urlArray` and `url` are maintained via
merging, the `loadstart` callback is still needed once, when the player
loads the initial video. See comment at the start of `init`.

Fixes #67